% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_name.R
\name{dynamic_name}
\alias{dynamic_name}
\title{Auto generate code to write out files with the date}
\usage{
dynamic_name(x)
}
\arguments{
\item{x}{character, scalar. A custom file name or Folder/in/Balitomore/name}
}
\value{
None
}
\description{
Auto generate code to write out files with the date
}
