% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorFix.R
\name{factorFix}
\alias{factorFix}
\title{Provides factor handling + mean imputation for numeric, etc}
\usage{
factorFix(train, test, exclude)
}
\arguments{
\item{train}{data.frame or similar classes}

\item{test}{data.frame or similar}

\item{exclude}{character scalar or vector of any column names to ignore}
}
\value{
None
}
\description{
Provides factor handling + mean imputation for numeric, etc
}
