% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmident.R
\name{rm_ident}
\alias{rm_ident}
\title{Removing identical features}
\usage{
rm_ident(train, test = NULL)
}
\arguments{
\item{train}{required. data.frame or similar classes}

\item{test}{optional. data.frame or similar classes}
}
\value{
None
}
\description{
Removing identical features
}
