% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff2.R
\name{setdiff2}
\alias{setdiff2}
\title{Look up student ID's using IP's (via SalesForce)}
\usage{
setdiff2(x, y)
}
\arguments{
\item{x}{data.frame or similiar}

\item{y}{data.frame or similiar}
}
\value{
None
}
\description{
Look up student ID's using IP's (via SalesForce)
}
