% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdesc.R
\name{stdesc}
\alias{stdesc}
\title{Fork of John C. Nash's function from Nonlinear Parameter Optimization using R Tools}
\usage{
stdesc(par, fn, gr, control = list(), ...)
}
\arguments{
\item{par}{required. a vector containing the starting point}

\item{fn}{required. objective function}

\item{gr}{required. gradient of objective function.}

\item{control}{optional. list of control parameters.}

\item{dowarn=TRUE}{by default. Set FALSE to suppress warnings}
}
\value{
ans  a list of results
}
\description{
Fork of John C. Nash's function from Nonlinear Parameter Optimization using R Tools
}
